Refactor routes and DB queries to account for userId from forwarded token

+++ src/plantTasks-database.js
@@ -34,69 +34,84 @@ class PlantTasksDatabase {
     return this.db.any('SELECT * FROM plants WHERE user_id = $1 ORDER BY id', userId);
   }
 
-  getPlant(plantId) {
-    return this.db.one('SELECT * FROM plants WHERE id = $1', plantId);
+  getPlant(plantId, userId) {
+    return this.db.one('SELECT * FROM plants WHERE id = $1 AND user_id = $2', [plantId, userId]);
   }
 
-  addPlant(plantName, plantSpecies, plantNotes) {
-    return this.db.one('INSERT INTO plants (name, species, notes) VALUES($1, $2, $3) RETURNING *',[plantName, plantSpecies, plantNotes]);
+  addPlant(plantName, plantSpecies, plantNotes, userId) {
+    return this.db.one('INSERT INTO plants (name, species, notes, user_id) VALUES($1, $2, $3, $4) RETURNING *',[plantName, plantSpecies, plantNotes, userId]);
   }
   
-  deletePlant(plantId) {
-    return this.db.one('DELETE FROM plants WHERE id = $1 RETURNING *', plantId);
+  deletePlant(plantId, userId) {
+    return this.db.one('DELETE FROM plants WHERE id = $1 AND user_id = $2 RETURNING *', [plantId, userId]);
   }
   
-  editPlant(plantId, plantName, plantSpecies, plantNotes) {
-    return this.db.one('UPDATE plants SET name = $2, species = $3, notes = $4 WHERE id = $1 RETURNING *', [plantId, plantName, plantSpecies, plantNotes]);
+  editPlant(plantId, plantName, plantSpecies, plantNotes, userId) {
+    return this.db.one('UPDATE plants SET name = $2, species = $3, notes = $4 WHERE id = $1 AND user_id = $5 RETURNING *', [plantId, plantName, plantSpecies, plantNotes, userId]);
   }
   
   // Tasks
-  getTaskOfPlant(plantId) {
+  getTaskOfPlant(plantId, userId) {
     return this.db.any(`SELECT t.*, p.name 
     FROM tasks AS t
     INNER JOIN plants AS p ON p.id = t.plant_id
-    WHERE p.id = $1`, plantId)
+    WHERE p.id = $1 AND p.user_id = $2`, [plantId, userId])
   }
   
-  getAllTasks() {
-    return this.db.any('SELECT * FROM tasks ORDER BY id')
+  getAllTasks(userId) {
+    return this.db.any(`SELECT t.* FROM tasks AS t 
+      INNER JOIN plants AS p ON p.id = t.plant_id
+      WHERE user_id = $1
+      ORDER BY t.id`, userId)
   }
   
-  addTask(description, frequency, plantId) {
-    return this.db.one('INSERT INTO tasks (description, frequency, plant_id) VALUES($1, $2, $3) RETURNING *', [description, frequency, plantId])
+  addTask(description, frequency, plantId, userId) {
+    return this.db.one(`INSERT INTO tasks (description, frequency, plant_id) 
+      SELECT ($1, $2, $3) 
+      WHERE EXISTS (SELECT * FROM plants WHERE user_id = $4)
+      RETURNING *`, [description, frequency, plantId, userId])
   }
   
-  deleteTask(taskId) {
-    return this.db.one('DELETE FROM tasks WHERE id = $1 RETURNING *', taskId);
+  deleteTask(taskId, userId) {
+    return this.db.one(`DELETE FROM tasks AS t 
+    USING plants AS p
+    WHERE t.id = $1 AND p.user_id = $2
+    RETURNING *`, [taskId, userId]);
   }
   
   // Task Instances
-  getTaskInstances() {
+  getTaskInstances(userId) {
     return this.db.any(`SELECT ti.id AS task_instance_id, ti.completed, ti.due_date, t.id AS task_id, t.description, t.frequency, p.id AS plant_id, p.name
       FROM task_instances AS ti 
       INNER JOIN tasks AS t ON t.id = ti.task_id
-      INNER JOIN plants AS p on p.id = t.plant_id`)
+      INNER JOIN plants AS p on p.id = t.plant_id
+      WHERE p.user_id = $1`, userId)
   }
 
   // convert UTC time to PST. temporary solution; will pass user's timezone later
-  getTodayTaskInstances() {
+  getTodayTaskInstances(userId) {
     return this.db.any(`SELECT ti.id AS task_instance_id, ti.completed, ti.due_date, t.id AS task_id, t.description, t.frequency, p.id AS plant_id, p.name
     FROM task_instances AS ti
     INNER JOIN tasks AS t ON t.id = ti.task_id
     INNER JOIN plants AS p on p.id = t.plant_id
-    WHERE date_trunc('day', ti.due_date) = date_trunc('day', now() AT TIME ZONE 'PST')`)
+    WHERE date_trunc('day', ti.due_date) = date_trunc('day', now() AT TIME ZONE 'PST') AND WHERE p.user_id = $1`, userId)
   }
   
-  getTaskInstancesByDay(date) {
+  getTaskInstancesByDay(date, userId) {
     return this.db.any(`SELECT ti.id AS task_instance_id, ti.completed, ti.due_date, t.id AS task_id, t.description, t.frequency, p.id AS plant_id, p.name
     FROM task_instances AS ti 
     INNER JOIN tasks AS t ON t.id = ti.task_id
     INNER JOIN plants AS p on p.id = t.plant_id 
-    WHERE date_trunc('day', ti.due_date) = $1)`, date)
+    WHERE date_trunc('day', ti.due_date) = $1) AND p.user_id = $2`, [date, userId])
   }
 
-  updateTaskInstance(status, taskInstanceId) {
-    return this.db.one('UPDATE task_instances SET completed = $1 WHERE id = $2 RETURNING *', [status, taskInstanceId])
+  updateTaskInstance(status, taskInstanceId, userId) {
+    return this.db.one(`UPDATE task_instances SET completed = $1 
+      WHERE id = (SELECT ti.id FROM task_instances AS ti 
+        INNER JOIN tasks AS t ON t.id = ti.task_id
+        INNER JOIN plants AS p on p.id = t.plant_id
+        WHERE ti.id = $2 AND p.user_id = $3)
+      RETURNING *`, [status, taskInstanceId, userId])
   }
 }
